requirements.txt
pyTelegramBotAPI==4.14.0
yt-dlp==2024.8.6
python-dotenv==1.0.1
